requirements.txt
setuptools
openpyxl >= 3.1.0
jinja2
pandas
pyyaml
